Use initptr_t under Win/mingw64 (sizeof(long)<sizeof(void*))

+++ src/OSD/OSD_WNT.cxx
@@ -1014,7 +1014,7 @@ BOOL MoveDirectory ( LPCTSTR oldDir, LPCTSTR newDir ) {
  DIR_RESPONSE        response;
  DWORD               level;
 
- if (   (  level = ( DWORD )TlsGetValue ( dwLevel )  ) == NULL   ) {
+ if (   (  level = ( intptr_t )TlsGetValue ( dwLevel )  ) == NULL   ) {
 
   ++level;
   TlsSetValue (  dwLevel, ( LPVOID )level  );
